fix android frontend connect to internal,gemini apis #38 Also began to split frontend source into model and view
