+++ .vscode/launch.json
@@ -1,0 +1,29 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "gemini_proact_flutter",
+            "cwd": "gemini_proact_flutter",
+            "request": "launch",
+            "type": "dart",
+            "program": "lib/view/main.dart"
+        },
+        {
+            "name": "gemini_proact_flutter (profile mode)",
+            "cwd": "gemini_proact_flutter",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "profile"
+        },
+        {
+            "name": "gemini_proact_flutter (release mode)",
+            "cwd": "gemini_proact_flutter",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "release"
+        }
+    ]
+}
